Fusion of config.hpp and config_impl.hpp
